feat: add split workspace file preview
