+++ packages/client/app/components/WorkspaceFilesPanel.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import { computed, nextTick, ref, watch } from 'vue'
-import { useLocalFileViewer } from '../composables/useLocalFileViewer'
+import { computed, ref, watch } from 'vue'
+import LocalFilePreview from './LocalFilePreview.vue'
 import {
   useWorkspaceFiles,
   type WorkspaceFileTreeNode
@@ -27,11 +27,8 @@ const {
   refreshCurrentDirectory,
   setShowIgnored
 } = useWorkspaceFiles(workspace)
-const { openViewer } = useLocalFileViewer()
-
 const open = ref(false)
 const copyStatus = ref<string | null>(null)
-const pendingViewerPath = ref<string | null>(null)
 
 const breadcrumbItems = computed(() => breadcrumbs.value.map((item, index) => ({
   ...item,
@@ -103,7 +100,7 @@ const handleTreeToggle = async (
   }
 }
 
-const handleTreeSelect = async (
+const handleTreeSelect = (
   _event: Event,
   node: WorkspaceFileTreeNode
 ) => {
@@ -113,26 +110,6 @@ const handleTreeSelect = async (
   }
 
   selectEntry(entry)
-  if (entry.kind !== 'file') {
-    return
-  }
-
-  pendingViewerPath.value = entry.path
-  open.value = false
-}
-
-const handleAfterLeave = async () => {
-  const path = pendingViewerPath.value
-  if (!path) {
-    return
-  }
-
-  pendingViewerPath.value = null
-  await nextTick()
-  openViewer({
-    workspace: props.workspace,
-    path
-  })
 }
 
 const handleSelectedNode = (node: WorkspaceFileTreeNode | undefined) => {
@@ -165,7 +142,6 @@ watch(
   () => {
     open.value = false
     copyStatus.value = null
-    pendingViewerPath.value = null
   }
 )
 
@@ -196,20 +172,18 @@ watch(open, (nextOpen) => {
     />
   </UTooltip>
 
-  <USlideover
+  <UModal
     :open="open"
     title="Workspace files"
     :description="`Read-only files rooted at ${workspaceLabel}.`"
-    side="right"
+    fullscreen
     dismissible
     :ui="{
-      content: 'w-[92vw] max-w-[92vw] sm:w-[28rem] sm:max-w-[28rem]',
+      content: 'overflow-hidden bg-default',
       header: 'px-4 py-3',
-      body: '!p-0 sm:!p-0',
-      footer: 'px-4 py-3'
+      body: 'min-h-0 overflow-hidden !p-0 sm:!p-0'
     }"
     @update:open="handleOpenChange"
-    @after:leave="handleAfterLeave"
   >
     <template #actions>
       <UTooltip text="Refresh current folder">
@@ -239,155 +213,182 @@ watch(open, (nextOpen) => {
     </template>
 
     <template #body>
-      <div class="flex h-full min-h-0 flex-col bg-default">
-        <div class="space-y-3 border-b border-default bg-elevated/20 px-4 py-3">
-          <UBreadcrumb
-            :items="breadcrumbItems"
-            class="min-w-0 overflow-x-auto"
-            :ui="{ list: 'flex-nowrap', linkLabel: 'max-w-32 truncate' }"
-          >
-            <template #item="{ item }">
-              <button
-                type="button"
-                class="max-w-32 truncate rounded text-sm hover:text-highlighted focus-visible:outline-2 focus-visible:outline-primary"
-                :aria-current="item.active ? 'page' : undefined"
-                @click="navigateTo(item.path)"
-              >
-                {{ item.label }}
-              </button>
-            </template>
-          </UBreadcrumb>
-
-          <div class="flex items-center justify-between gap-3">
-            <UCheckbox
-              :model-value="snapshot.showIgnored"
-              label="Show generated folders"
-              size="sm"
-              @update:model-value="(value: boolean | 'indeterminate') => setShowIgnored(Boolean(value))"
-            />
-            <span class="text-xs text-muted">Read only</span>
-          </div>
-
-          <p
-            v-if="copyStatus"
-            class="text-xs text-muted"
-            role="status"
-          >
-            {{ copyStatus }}
-          </p>
-        </div>
-
-        <UScrollArea class="min-h-0 flex-1 px-2 py-3">
-          <div
-            v-if="rootLoading"
-            class="space-y-2 px-2"
-            aria-label="Loading workspace files"
-          >
-            <USkeleton
-              v-for="index in 6"
-              :key="index"
-              class="h-8 w-full"
-            />
-          </div>
-
-          <div
-            v-else-if="rootError"
-            class="space-y-3 px-2"
-          >
-            <UAlert
-              color="error"
-              variant="soft"
-              icon="i-lucide-triangle-alert"
-              title="Could not load workspace files"
-              :description="rootError"
-            />
-            <UButton
-              label="Retry"
-              icon="i-lucide-refresh-cw"
-              color="neutral"
-              variant="outline"
-              size="sm"
-              @click="loadDirectory('', { force: true })"
-            />
-          </div>
-
-          <div
-            v-else-if="currentDirectoryError"
-            class="space-y-3 px-2 pb-3"
-            role="status"
-            aria-live="polite"
-          >
-            <UAlert
-              color="error"
-              variant="soft"
-              icon="i-lucide-triangle-alert"
-              title="Could not load folder"
-              :description="currentDirectoryError"
-            />
-            <UButton
-              label="Retry folder"
-              icon="i-lucide-refresh-cw"
-              color="neutral"
-              variant="outline"
-              size="sm"
-              @click="loadDirectory(snapshot.currentPath, { force: true })"
-            />
-          </div>
-
-          <UTree
-            v-if="!rootLoading && !rootError"
-            :items="treeItems"
-            :get-key="(item: WorkspaceFileTreeNode) => item.key"
-            :model-value="selectedNode"
-            :expanded="snapshot.expandedPaths"
-            color="primary"
-            size="sm"
-            class="w-full"
-            :ui="{
-              link: 'min-w-0 justify-start rounded-lg text-left',
-              linkLabel: 'min-w-0 flex-1 text-left',
-              linkLeadingIcon: 'shrink-0'
-            }"
-            @update:model-value="handleSelectedNode"
-            @update:expanded="(paths: string[]) => { snapshot.expandedPaths = paths }"
-            @toggle="handleTreeToggle"
-            @select="handleTreeSelect"
-          >
-            <template #item-label="{ item }">
-              <span
-                class="flex min-w-0 w-full items-center justify-start gap-2 text-left"
-                :class="item.status ? 'text-muted' : ''"
-              >
-                <span class="min-w-0 flex-1 truncate">{{ item.label }}</span>
-                <UBadge
-                  v-if="item.entry?.isSymlink"
-                  color="neutral"
-                  variant="subtle"
-                  size="xs"
+      <div class="grid h-full min-h-0 grid-rows-[minmax(14rem,42%)_minmax(0,1fr)] bg-default md:grid-cols-[minmax(18rem,26rem)_minmax(0,1fr)] md:grid-rows-1">
+        <aside class="flex min-h-0 flex-col border-b border-default bg-default md:border-r md:border-b-0">
+          <div class="space-y-3 border-b border-default bg-elevated/20 px-4 py-3">
+            <UBreadcrumb
+              :items="breadcrumbItems"
+              class="min-w-0 overflow-x-auto"
+              :ui="{ list: 'flex-nowrap', linkLabel: 'max-w-32 truncate' }"
+            >
+              <template #item="{ item }">
+                <button
+                  type="button"
+                  class="max-w-32 truncate rounded text-sm hover:text-highlighted focus-visible:outline-2 focus-visible:outline-primary"
+                  :aria-current="item.active ? 'page' : undefined"
+                  @click="navigateTo(item.path)"
                 >
-                  link
-                </UBadge>
+                  {{ item.label }}
+                </button>
+              </template>
+            </UBreadcrumb>
+
+            <div class="flex items-center justify-between gap-3">
+              <UCheckbox
+                :model-value="snapshot.showIgnored"
+                label="Show generated folders"
+                size="sm"
+                @update:model-value="(value: boolean | 'indeterminate') => setShowIgnored(Boolean(value))"
+              />
+              <span class="text-xs text-muted">Read only</span>
+            </div>
+
+            <p
+              v-if="copyStatus"
+              class="text-xs text-muted"
+              role="status"
+            >
+              {{ copyStatus }}
+            </p>
+          </div>
+
+          <UScrollArea class="min-h-0 flex-1 px-2 py-3">
+            <div
+              v-if="rootLoading"
+              class="space-y-2 px-2"
+              aria-label="Loading workspace files"
+            >
+              <USkeleton
+                v-for="index in 6"
+                :key="index"
+                class="h-8 w-full"
+              />
+            </div>
+
+            <div
+              v-else-if="rootError"
+              class="space-y-3 px-2"
+            >
+              <UAlert
+                color="error"
+                variant="soft"
+                icon="i-lucide-triangle-alert"
+                title="Could not load workspace files"
+                :description="rootError"
+              />
+              <UButton
+                label="Retry"
+                icon="i-lucide-refresh-cw"
+                color="neutral"
+                variant="outline"
+                size="sm"
+                @click="loadDirectory('', { force: true })"
+              />
+            </div>
+
+            <div
+              v-else-if="currentDirectoryError"
+              class="space-y-3 px-2 pb-3"
+              role="status"
+              aria-live="polite"
+            >
+              <UAlert
+                color="error"
+                variant="soft"
+                icon="i-lucide-triangle-alert"
+                title="Could not load folder"
+                :description="currentDirectoryError"
+              />
+              <UButton
+                label="Retry folder"
+                icon="i-lucide-refresh-cw"
+                color="neutral"
+                variant="outline"
+                size="sm"
+                @click="loadDirectory(snapshot.currentPath, { force: true })"
+              />
+            </div>
+
+            <UTree
+              v-if="!rootLoading && !rootError"
+              :items="treeItems"
+              :get-key="(item: WorkspaceFileTreeNode) => item.key"
+              :model-value="selectedNode"
+              :expanded="snapshot.expandedPaths"
+              color="primary"
+              size="sm"
+              class="w-full"
+              :ui="{
+                link: 'min-w-0 justify-start rounded-lg text-left',
+                linkLabel: 'min-w-0 flex-1 text-left',
+                linkLeadingIcon: 'shrink-0'
+              }"
+              @update:model-value="handleSelectedNode"
+              @update:expanded="(paths: string[]) => { snapshot.expandedPaths = paths }"
+              @toggle="handleTreeToggle"
+              @select="handleTreeSelect"
+            >
+              <template #item-label="{ item }">
                 <span
-                  v-if="item.entry?.kind === 'file' && item.entry.size !== null"
-                  class="shrink-0 text-[11px] text-dimmed"
+                  class="flex min-w-0 w-full items-center justify-start gap-2 text-left"
+                  :class="item.status ? 'text-muted' : ''"
                 >
-                  {{ formatLocalFileSize(item.entry.size) }}
+                  <span class="min-w-0 flex-1 truncate">{{ item.label }}</span>
+                  <UBadge
+                    v-if="item.entry?.isSymlink"
+                    color="neutral"
+                    variant="subtle"
+                    size="xs"
+                  >
+                    link
+                  </UBadge>
+                  <span
+                    v-if="item.entry?.kind === 'file' && item.entry.size !== null"
+                    class="shrink-0 text-[11px] text-dimmed"
+                  >
+                    {{ formatLocalFileSize(item.entry.size) }}
+                  </span>
+                  <span
+                    v-if="entryStatusLabel(item)"
+                    class="shrink-0 text-[11px] text-warning"
+                  >
+                    {{ entryStatusLabel(item) }}
+                  </span>
                 </span>
-                <span
-                  v-if="entryStatusLabel(item)"
-                  class="shrink-0 text-[11px] text-warning"
-                >
-                  {{ entryStatusLabel(item) }}
-                </span>
-              </span>
-            </template>
-          </UTree>
-        </UScrollArea>
+              </template>
+            </UTree>
+          </UScrollArea>
 
-        <div class="border-t border-default px-4 py-3 text-xs leading-5 text-muted">
-          Select a supported text or image file to open the existing preview. File changes are not available here.
-        </div>
+          <div class="border-t border-default px-4 py-3 text-xs leading-5 text-muted">
+            Read-only workspace files. File changes are not available here.
+          </div>
+        </aside>
+
+        <section
+          class="min-h-0 bg-elevated/10"
+          aria-label="Workspace file preview"
+        >
+          <LocalFilePreview
+            v-if="selectedEntry?.kind === 'file' && selectedEntry.accessible"
+            :workspace="workspace"
+            :path="selectedEntry.path"
+            eyebrow="Workspace file preview"
+          />
+          <div
+            v-else
+            class="flex h-full min-h-0 flex-col items-center justify-center gap-4 px-6 py-10 text-center"
+          >
+            <UIcon
+              name="i-lucide-file-search-2"
+              class="size-12 text-muted/45"
+              aria-hidden="true"
+            />
+            <p class="max-w-sm text-sm text-muted">
+              Select a file from the tree to preview it.
+            </p>
+          </div>
+        </section>
       </div>
     </template>
-  </USlideover>
+  </UModal>
 </template>
